autofs-5.1.3 - workaround getaddrinfo(3) ai_canonname bug
There appears to be a bug in some versions of getaddrinfo() where the
addrinfo structure gets garbage in ai_canonname when it is not requested
by the lookup flags causing a crash or lookup fail.

Try and work around this by always requesting ai_canonname for lookups
and hope this doesn't affect the semantics of the autofs lookup usage.

Signed-off-by: Ian Kent <raven@themaw.net>