autofs-5.1.2 - use autofs_point to store expire timeout where possibe
For technical reasons the expire timeout is stored in the map entry source
rather than the struct autofs_point and must be stored in the map source for
direct (and offset) mounts.

But with the map source re-use needed by amd format type "auto" mounts the
timeout can't be stored in the map source since the it may be different for
mounts that share the map source.

So use the struct autofs_point where possible.

Signed-off-by: Ian Kent <raven@themaw.net>