autofs-5.1.2 - add sss master map wait config option
When sssd is starting up it can sometimes return "no such entry" for a
short time until it has read in the LDAP map information. This affects
reading the master map at autofs startup and results in no automount
mounts if sssd is the primary map source.

This problem should be resolved in sssd but it's a problem for the
moment so a configuration option, sss_master_map_wait, has been added
to work around it.

The internal program default is 0, don't wait. If the work around is
needed try setting this to 10 seconds to work around it.

Signed-off-by: Ian Kent <raven@themaw.net>