autofs-5.1.2 - check NFS server availability on local mount fallback
The availability probe isn't done for anything autofs thinks is a
local mount because it's the local machine.

It first tries a bind mount and if that fails it falls back to trying
a local NFS mount. If the local NFS server is not running mount.nfs(8)
can suffer a lengthy timeout.

So check for the bind mount fallback case and check if an NFS server
is responding before trying the mount.

Signed-off-by: Ian Kent <raven@themaw.net>