autofs-5.1.2 - fix offset mount location multiple expansion
When parsing an offset map entry mount location it shouldn't be
expanded because it will have already been dequoted before being
added to the map entry cache when it was first parsed.

Expanding it again can lead to inconsistent behaviour between
the expansion of simple map entry locations and offset locations
that have had quoting removed as part of the initial parse.

Signed-off-by: Ian Kent <raven@themaw.net>