autofs-5.1.2 - increase worker thread per-thread stack size
automount(8) uses a worker thread per-thread stack size of 1M which is
sufficient for its needs.

But some glibc functions (such as nscd_getgr_r()) use alloca() to allocate
working storage that can have an arbitary size. Since alloca() never fails
there is no way to check for stack overflow. This has been fixed in more
recent versions of glibc.

But for older version of glibc all that can be done by automount to avoid
this is to increase the stack size.

Signed-off-by: Ian Kent <raven@themaw.net>