autofs-5.1.2 - add configure option for limiting getgrgid_r() stack usage
Almost all the time it isn't a problem for glibc to use stack allocation
to store group information during calls to getgrgid_r().

But if it is a problem the --enable-limit-getgrgid-size configure option
can be used to limit the buffer size passed to getgrgid_r() (which is used
to decide whether to alloca() local storage for the call).

The check allows the call to go ahead if the increased buffer size is less
than 0.9 of the thread stack size.

This isn't ideal because the current stack usage isn't known but should
be ok much of the time.

Signed-off-by: Ian Kent <raven@themaw.net>