autofs-5.1.3 - Add -c option when calling /bin/umount - if supported.
The "-c" option has been supported by umount since util-linux 2.17.

It tells umount that the path name is "canonical", meaning no symlinks
or ".." etc.

This is appropriate for autofs to use as it always uses canonical path
names. The advantage of "-c" is that it means umount doesn't need to
'lstat()' the path so much.

From util-linux 2.30, umount doesn't lstat the path at all when "-c"
is passed. This is particularly beneficial when unmounting an NFS
filesystem for which the server is no reachable. In that case an
'lstat()' will hang, but the umount(2) can succeed.

So check if "-c" is supported, and if so, use it.
Prior to 2.30 this doesn't help much, but doesn't hurt.
After 2.30 (and a similar small change to nfs-utils), "-c" will mean
that unmounting NFS filesystems will not hang.

Signed-off-by: NeilBrown <neilb@suse.com>
Signed-off-by: Ian Kent <raven@themaw.net>