autofs-5.1.3 - fix mount.nfs blocks on first mount
The first time an NFS mount is tried mount.nfs tries to start rpc.statd
using systemctl. If /usr/local is being used to provide automounts in
that directory the first mount will hang becuase some directories under
/usr/local are included in the path used by systemctl.

Add rpc-statd.service (and for good measure rpcbind.service) to the
autofs Wants line in the unit file to work around this.

Signed-off-by: Ian Kent <raven@themaw.net>