autofs-5.1.3 - port option should not behave like nobind option
The "port=" option currently prevents bind mounting when the host
corresponds to localhost or a local interface.

But this interferes with cases where the "port" option needs to be
used and bind mounting of local mounts is also required.

So remove the age old hack of the "port" option preventing this in
favour of explicitly requiring the "nobind" option.

Signed-off-by: Ian Kent <raven@themaw.net>