autofs-5.1.3 - fix amd defaults map entry handling
When parsing an amd map entry the map entry defaults were not being
handled properly.

A "-" in the map entry requires a reset to defaults, which wasn't being
done.

A "-<map options>" in the map entry requires merging the <map options>
with the current defaults map entry, which was also not being done
properly.

Signed-off-by: Ian Kent <raven@themaw.net>