autofs-5.1.3 - reset master map list on startup retry
On autofs startup if automount thinks the master map read failed
for some reason it retries the read after waiting for a fairly
short delay.

It should be ok to retry with the entries seen so far but there
may be cases where that causes problems so release any existing
master map entries before retrying.

Signed-off-by: Ian Kent <raven@themaw.net>