autofs-5.1.3 - move open_xxxx() functions to spawn.c
In a bug report by John Salmon has described a race between the
autofs open_xxxx() functions and fork(2) that can cause file handles
that don't have close on exec set to leak into subprocesses.

Basically, in some cases there can be a finite time between performing
the open and setting the descriptor close on exec and it's this window
that leads to the problem.

The description went on to talk about a case where autofs can hang when
this happens. That is when the leaked decriptor causes poll(2) to not
return in another process because it is waiting for the decriptor to
close (on mount completion) but another execed process has the cloned
descriptor open.

Serializing access to the open_xxxx() functions wrt. to fork(2) calls
should be suficient to resolve this leakage.

This patch starts this by moving the open_xxxx() out of the automount.h
include file and into spawn.c.

Signed-off-by: Ian Kent <raven@themaw.net>