autofs-5.1.4 - fix use after free in do_master_list_reset()
Umm ... list_for_each() can't be used in do_master_list_reset() because
the subject entry of the loop is removed for the list within the loop
body. Therefore it can't be used to calculate the next pointer within a
for (...) loop.

There is no list_for_each_safe() macro in the list.h of autofs so it
needs to be done manually.

Signed-off-by: Ian Kent <raven@themaw.net>