autofs-5.1.4 - account for recent libnsl changes
The glibc RPC code has been dropped.

The NIS functionality comes in the libnsl library but it installs
its files into sub directories of the system include and library
directories.

So configure needs to check for this and allow for it in the
compile and linking of the NIS and NIS+ lookup modules.