autofs-5.1.4 - tiny patch for autofs typo and possible bug
(1) The word "to" is doubled in two warnings.

(2) It prints a warning when it can't open the "old" config file, even
if it was able to open the default one. That's a reasonable warning if
it *couldn't* open the new one, but seems unnecessary if it could.

The patch removes the unneeded "to"s and only prints warning about the
old config file if it couldn't open the default one.

Signed-off-by: Todd Eigenschink <todd@fai2.com>
Signed-off-by: Ian Kent <raven@themaw.net>