autofs-5.1.4 - use systemd sd_notify() at startup
autofs needs to ensure statd is started before any NFS mounts
are attempted.

When starting the statd service with the autofs service the statd
service will trigger a restart of the autofs service during its
start up. Sometimes this can happen during the automount start up
itself.

When this happens it causes systemd to become confused and remove
the pid file created by automount leaving systemd thinking the
autofs service had failed start up when it was actually running.

It was recommened that autofs be changed to a "Type=notify" service
to avoid this. Using this a pid file is no longer needed and is now
not used.

Signed-off-by: Ian Kent <raven@themaw.net>