autofs-5.1.4 - add units After line to include statd service
autofs needs to ensure statd is started before any NFS mounts
are attempted. This is because if the first NFS mount starts
the service and the mount is an automounted directory that is
included in the PATH used by systemctl (eg. /usr/local/bin)
the mount cannot complete.

Add rpc-statd.service to the unit "After=" line to try and
ensure it is started before automount.

Signed-off-by: Ian Kent <raven@themaw.net>