autofs-5.1.4 - enable SIGUSR2 handling in rpm spec file
Because the preferred usage is to leave busy mounts in place and
re-connect to them on start up the SIGUSR2 logic hasn't been used
or tested for a long time.

But it can be useful so enable it so any problems with it can be
identified and resolved.

It turns out that the SIGUSR2 forced shutdown behaves like a SIGUSR1
prune expire excpet that autofs will shutdown if all current mounts
are expired. But the kernel currently won't trigger expires for
mounts that are in use so it doesn't function quite as intended.

Enable the forced shutdown anyway so it will function if the kernel
supports it.

Signed-off-by: Ian Kent <raven@themaw.net>