autofs-5.1.4 - use defines for expire type
The kernel defines for expire type such as an immediate expire
shoule be used to clearify what is being requested.

AUTOFS_EXP_IMMEDIATE corresponds to a SIGUSR1 prune operation.

AUTOFS_EXP_FORCE corresponds to an expire type not yet implemented in
the kernel, a SIGUSR2 forced expire. Define it in our internal autofs
kernel include file, the kernel will ignore it if it doesn't support it.

AUTOFS_EXP_LEAVES is no longer used in autofs version 5.

Finally add a define AUTOFS_EXP_NORMAL to indicate we're perfoming a
normal expire.

Signed-off-by: Ian Kent <raven@themaw.net>