autofs-5.1.4 - fix age setting at startup
Commit 2b567ace7d, which resets the master map list on startup
before retrying a read, incorrectly sets age using time() when
it should use monotonic_time().

This causes lookup failures for submounts in cases where a master
map read retry was needed.

The failure happens because the mount entry age is greater than
the map entry age which is meant to indicate the map is no longer
valid which is not the case.

Signed-off-by: Ian Kent <raven@themaw.net>