autofs-5.1.4 - add master map pseudo options for mount propagation
Add master map entry pseudo mount option of "slave" or "private" to
allow mount propagation of bind mounts to be set to either "slave" or
"private".

This option may be needed when using multi-mounts that have bind mounts
that bind to a file system that is propagation shared. This is becuase
the bind mount will have the same properties as its target which causes
problems for offset mounts. When this happens an unwanted offset mount
is propagated back to the target file system resulting in a deadlock
when attempting to access the offset.

By default bind mounts will inherit the mount propagation of the target
file system.

Signed-off-by: Ian Kent <raven@themaw.net>