autofs-5.1.4 - fix update_negative_cache() map source usage
File map sources can be either plain text or executable.

When the map path is specified without a type (eg. when a
full path is used) an instance map source is used and the
original map is left unchanged.

But update_negative_cache() fails to take this into account
causing it to update the wrong map cache.

When a map reload is done the map entry appears to not exist
so the new map entry is added.

This could go unnoticed except that, after a map read, the
map entry cache cleans stale map entries and the existence
of this negative entry causes the new map entry to be deleted
and map lookups continue to fail.

In hindsite the use of an instance map source for this is
probably uneccessary but changing it will be risky so, for
now, just make update_negative_cache() use the correct map.

Signed-off-by: Ian Kent <raven@themaw.net>