autofs-5.1.4 - mark removed cache entry negative
When re-reading a map, entries that have been removed are detected
and deleted from the map entry cache by lookup_prune_cache().

If a removed map entry is mounted at the time lookup_prune_cache()
is called the map entry is skipped. This is done becuase the next
lookup (following the mount expire, which needs the cache entry to
remain) will detect the stale cache entry and a map update done
resulting in the stale entry being removed.

But if a map re-read is performed while the cache entry is mounted
the cache will appear to up to date so the removed entry will remain
valid even after it has expired.

To cover this case it's sufficient to mark the mounted cache entry
negative during the cache prune which prevents further lookups from
using the stale entry.

Signed-off-by: Ian Kent <raven@themaw.net>