autofs-5.1.4 - add systemd service command line option
When run as a systemd service using the systemd notification method to
synchronise startup, logging should be done to syslog so the log entry
format is consistent between daemon and systemd usage.

So, rather than run use the forground option, add an option to tell
the automounter it's being run as a systemd service and use syslog
for logging when its present on the command line.

Signed-off-by: Ian Kent <raven@themaw.net>