autofs-5.1.5 - update autofs(8) offset map entry update description
The autofs(8) man page imples that map entries that have offsets will
be updated in the ame way as top level map keys.

This isn't correct, if a map entry has offsets and is in use when the
map is updated the mount must expire before it will see the update.

Signed-off-by: Ian Kent <raven@themaw.net>