autofs-5.1.5 - always use PROC_MOUNTS to make mount lists
If it's necessary to read a mount table then always use the proc file
system mount table.

This could be very inefficient for cases where the mtab (_PATH_MOUNTED)
is specified but the mtab has been a symlink to the proc mount tables
for quite a while now so it doesn't make any difference always using
the proc tables.

Signed-off-by: Ian Kent <raven@themaw.net>