autofs-5.1.5 - don't use tree_is_mounted() for mounted checks
Always use is_mounted() for mounted checks.

If the proc mount table actually needs to be used this could be
very inefficient but the miscelaneous device ioctl interface
has been available to quite a few years now so it really needs
to be used to avoid the scanning of mount tables.

Signed-off-by: Ian Kent <raven@themaw.net>