autofs-5.1.5 - fix unlink_mount_tree() not umounting mounts
At some point a check matching process group id of mounts was added to
the unlink_mount_tree() function so that mounts not matching the group
id of the calling process would be skipped.

But unlink_mount_tree() is meant to force a cleanup of mounts from a
previous automount process so the process group of the current process
can never match so all mounts are skipped.

Signed-off-by: Ian Kent <raven@themaw.net>