autofs-5.1.5 - add config option for "ignore" mount option
Add a configuration option to control whether the autofs pseudo
mount option is used on autofs mounts.

The default setting is "no" to avoid unexpected behaviour and
so is an opt-in setting for those who understand that, if user
space utilities and libraries honour this, then autofs mounts
will be ommitted from mount table listings.

Signed-off-by: Ian Kent <raven@themaw.net>