autofs-5.1.5 - refactor unlink_active_mounts() in direct_c
In daemon/direct.c:unlink_active_mounts() the tree_get_mnt_list() call
returns 1 for a non-empty list which amounts to a mounted check and then
handles the case where the map is being read.

But the check that's carried out is to distinguish between a readmap
and a mount activity which should be handled by the caller instead of
burying the check away in the unlink_active_mounts() function.

Signed-off-by: Ian Kent <raven@themaw.net>