autofs-5.1.6 - update ldap READMEs and schema definitions
The autofs schema in samples/autofs.schema should not be used for
autofs map information, it's very old and may be inaccurate or may
conflict with other schema definitions included in LDAP server
distributions.

The README.autofs-schema has been updated to alert people to this
but the schema file has not yet been removed.

A new README.ldap-schema has been added which recommends using either
of rfc2307 or rfc2307bis schema for autofs Sun format map information
stored in LDAP and at least one of these schema should be included in
LDAP server distributions. Additionally the README notes the schema
that needs to be used for autofs amd format maps is present in the
file samples/am-utils-ldap.schema.

Ian

Signed-off-by: Ian Kent <raven@themaw.net>