autofs-5.1.6 - fix a regression with map instance lookup
Commit b66deff4241d ("autofs-5.1.3 - fix possible map instance memory
leak") introduced a regression.

The change didn't fix the memory leak and also failed to fix the race
updating the map instance list that caused it.

To fix this get rid of the horible temporary map usage and update the
instance list in place. Doing this causes some additional allocations
and frees but is somewhat simpler overall and avoids the race.

Fixes: b66deff4241d ("autofs-5.1.3 - fix possible map instance memory leak")
Signed-off-by: Ian Kent <raven@themaw.net>