autofs-5.1.6 - fix ldap sasl reconnect problem
When performing an ldap sasl connection a two step initialisation
was being done in an attempt to partially reuse existing connection
setup.

But if a network connectivity problem occurs the connection can end
up only half initialized and recovery after connectivity is restored
fails.

So get rid of the two step initialization, as it's benefit was at best
questionable, so that connection attempts either succeed or completely
fail. This leaves the connection completely uninitialized if there's a
network conectivity problem, ready for a new connection attempt.

Signed-off-by: Ian Kent <raven@themaw.net>