autofs-5.1.6 - samples/ldap.schema fix
This bug starts with version 5.1.3, and continues up to and including
the latest 5.1.6.  Version 5.1.2 was fine.

When slapd is invoked while including the sample LDAP schema file
autofs.schema, slapd will crash with the following error:

5f359370 /<path-to>/autofs.schema: line 14 attributetype:
AttributeType inappropriate matching rule: "caseExactMatch"

The problem is on line 13, which reads:

  EQUALITY caseExactMatch

It should read:

  EQUALITY caseExactIA5Match

Attached is a patch that will make the necessary change to
samples/autofs.schema that works for all versions 5.1.3 to 5.1.6.

Signed-off-by: Michael Peek <peek@nimbios.org>
Signed-off-by: Ian Kent <raven@themaw.net>