autofs: Fix crash in sun_mount()
In sun_mount() the variable np gets initialized to an alloca()
memory area:
  np = noptions = alloca();
Later on, at the end of a loop, it may get accessed like this:
  *(np - 1) = '\0';

If np hasn't been increased in between those lines, this access triggers
an out-of-bounds access which overwrites stack area and on the parisc
architecture segfaults the automount executable as described in the
Debian bugzilla #892953.

The patch below adds the necessary check and thus fixes the crash.

Bug: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=892953

Signed-off-by: Helge Deller <deller@gmx.de>
Reviewed-by: Paul Menzel <pmenzel@molgen.mpg.de>
Signed-off-by: Ian Kent <raven@themaw.net>