autofs-5.1.6 - fix unlink mounts umount order
The recent changes to mount table handling to support the "ignore"
autofs pseudo option resulted in the incorrect umount order being used
in the unlink_mount_tree() function.

To fix this change unlink_mount_tree() to use the existing get_mnt_list()
function to construct a correctly ordered list for the umounts.

Signed-off-by: Ian Kent <raven@themaw.net>